Tests after migration to SPQR #139
Refactoring tests for dates
