+++ graphql/src/test/resources/graphql/io/jmix/graphql/datafetcher/query-datatypesTestEntity-dates-with-filter.graphql
@@ -1,0 +1,14 @@
+query($filter:inp_scr_DatatypesTestEntityFilterCondition){
+  scr_DatatypesTestEntityList(
+    filter: [$filter]
+  ){
+    dateAttr
+    timeAttr
+    dateTimeAttr
+    localDateAttr
+    localTimeAttr
+    localDateTimeAttr
+    offsetDateTimeAttr
+    offsetTimeAttr
+  }
+}
